+++ app/api/api/signals.py
@@ -4,7 +4,7 @@ from django.core.cache import cache
 from django.db.models import Model
 from django.db.models.signals import post_delete, post_save
 from django.dispatch import receiver
-from shared.mapping.models import Dataset, Project, ScanReport
+from mapping.models import Dataset, Project, ScanReport
 
 
 @receiver(post_save, sender=Project)
