+++ app/api/config/settings.py
@@ -66,8 +66,7 @@ INSTALLED_APPS = [
     "datasets",
     "projects",
     "api",
-    "shared.data",
-    "shared.mapping",
+    "mapping",
     "rest_framework",
     "django_filters",
     "rest_framework.authtoken",
