+++ app/api/mapping/apps.py
@@ -2,5 +2,5 @@ from django.apps import AppConfig
 
 
 class MappingConfig(AppConfig):
-    name = "shared.mapping"
+    name = "mapping"
     label = "mapping"
