+++ app/api/mapping/management/commands/add_datasets_to_partner.py
@@ -1,5 +1,5 @@
 from django.core.management.base import BaseCommand
-from shared.mapping.models import DataPartner, Dataset
+from mapping.models import DataPartner, Dataset
 
 
 class Command(BaseCommand):
