+++ app/api/mapping/management/commands/airflow_schema_creation.py
