+++ app/api/mapping/management/commands/automatic_queue_and_containers_creation.py
