+++ app/api/mapping/management/commands/automatic_seeding_data.py
@@ -1,6 +1,6 @@
 from django.core.management.base import BaseCommand
 from django.core.management import call_command
-from shared.mapping.models import OmopTable
+from mapping.models import OmopTable
 from shared.files.models import FileType
