+++ app/api/mapping/management/commands/generate_rules_json.py
@@ -1,7 +1,7 @@
 import json
 
 from django.core.management.base import BaseCommand
-from shared.mapping.models import MappingRule
+from mapping.models import MappingRule
 from shared.services.rules_export import get_mapping_rules_json, make_dag
