+++ app/api/mapping/management/commands/unorphan_scanreports.py
@@ -1,5 +1,5 @@
 from django.core.management.base import BaseCommand
-from shared.mapping.models import Dataset, Project, ScanReport
+from mapping.models import Dataset, Project, ScanReport
 
 
 class Command(BaseCommand):
