+++ app/api/mapping/permissions.py
@@ -5,7 +5,7 @@ from django.contrib.auth.models import User
 from django.db.models.query_utils import Q
 from rest_framework import permissions
 from rest_framework.request import Request
-from shared.mapping.models import (
+from mapping.models import (
     Dataset,
     ScanReport,
     ScanReportField,
