+++ app/api/projects/serializers.py
@@ -1,6 +1,6 @@
 from drf_dynamic_fields import DynamicFieldsMixin  # type: ignore
 from rest_framework import serializers
-from shared.mapping.models import Project
+from mapping.models import Project
 from shared.users.serializers import UserSerializer
