+++ app/api/test/test_permissions.py
@@ -7,14 +7,14 @@ from projects.views import ProjectDetail
 from rest_framework.authtoken.models import Token
 from rest_framework.generics import GenericAPIView
 from rest_framework.test import APIRequestFactory, force_authenticate
-from shared.mapping.models import (
+from mapping.models import (
     DataPartner,
     Dataset,
     Project,
     ScanReport,
     VisibilityChoices,
 )
-from shared.mapping.permissions import (
+from mapping.permissions import (
     CanAdmin,
     CanEdit,
     CanView,
