+++ app/api/test/test_services_rules.py
@@ -1,12 +1,9 @@
-from datetime import date
-
-import pytest
 from django.contrib.auth import get_user_model
 from django.contrib.contenttypes.models import ContentType
 from django.test import TestCase
 from rest_framework.authtoken.models import Token
 from data.models import Concept
-from shared.mapping.models import (
+from mapping.models import (
     DataPartner,
     Dataset,
     MappingRule,
